honest wire widths: Remapper at the JS exporters, overrides and the JsonValue bridge gone (#156)
The contract now declares true Rust widths - i64 ms epochs and counts,
serde_json::Value for opaque JSON - with no per-field #[specta(type = ...)]
narrowing. The JS-facing exporters apply specta-util's
dangerous_bigints_as_number at export instead (vegify-typegen for the TS
artifact; the desktop's ttipc bindings via the new Bindings::map_datatypes
seam, ttipc branch rev 99f7fcf), which is the remapper's documented use: the
narrowing is invisible in TypeScript, and schema exporters keep the truth -
openapi.json now says type: integer with full i64 bounds where it said number.

Fallout, all of it good: the JsonValue bridge enum is deleted (specta's own
serde_json::Value export replaces it; the TS type is Value), the generated SDK
types carry honest i64s, and the old f64 declarations' NaN-nullability is gone
so bindings.ts tightens lastAt/unread/createdAt from number | null to number.
The bell's Notification moves to a hand-written wire struct in the SDK: its
payload hits specta-rust's anonymous-inline-enum limit
(serde_json::Value.Number.0 - the exporter errors loudly, as designed), so it
waits out upstream rendering Value as an opaque; field idents already follow
the generated convention, so the eventual move is a deletion. With no untagged
enum left in the generated slice, the untagged patch in typegen is gone too.

Verified: full just check green; live prod pull deserializes through the
generated i64 types; openapi.json 3.0.3 with 20 paths intact.

+++ apps/desktop/src-tauri/src/data.rs
@@ -264,7 +264,7 @@ fn apply_pull(
         if let Some(ts) = ing.deletedAt {
             tx.execute(
                 "UPDATE ingredients SET deleted_at = ?1 WHERE id = ?2",
-                params![ts as i64, ing.id],
+                params![ts, ing.id],
             )?;
         }
     }
